split bundler stuff out (naively)
