+++ src/index.ts
@@ -5,61 +5,17 @@ import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 import { ContractSuite } from "./safe";
 import { getNearSignature } from "./near";
+import {
+  getUserOpReceipt,
+  packPaymasterData,
+  sendUserOperation,
+} from "./bundler";
 
 dotenv.config();
 const { SAFE_SALT_NONCE, ERC4337_BUNDLER_URL, ETH_RPC, RECOVERY_ADDRESS } =
   process.env;
 const DUMMY_ECDSA_SIG =
   "0x000000000000000000000000ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff";
-
-async function sendUserOperation(userOp: UserOperation, entryPoint: string) {
-  const response = await fetch(ERC4337_BUNDLER_URL!, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      jsonrpc: "2.0",
-      method: "eth_sendUserOperation",
-      id: 4337,
-      params: [userOp, entryPoint],
-    }),
-  });
-  const body = await response.text();
-  if (!response.ok) {
-    throw new Error(`Failed to send user op ${body}`);
-  }
-  const json = JSON.parse(body);
-  if (json.error) {
-    throw new Error(JSON.stringify(json.error));
-  }
-  // This does not contain a transaction receipt! It is the `userOpHash`
-  return json.result;
-}
-
-async function getUserOpReceipt(userOpHash: string) {
-  const response = await fetch(ERC4337_BUNDLER_URL!, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      jsonrpc: "2.0",
-      method: "eth_getUserOperationReceipt",
-      id: 4337,
-      params: [userOpHash],
-    }),
-  });
-  const body = await response.text();
-  if (!response.ok) {
-    throw new Error(`Failed to send user op ${body}`);
-  }
-  const json = JSON.parse(body);
-  if (json.error) {
-    throw new Error(JSON.stringify(json.error));
-  }
-  return json.result;
-}
 
 async function main() {
   const argv = await yargs(hideBin(process.argv)).option("usePaymaster", {
@@ -175,6 +131,7 @@ async function main() {
     [0, 0, await getNearSignature(nearAdapter, safeOpHash)],
   );
   const userOpHash = await sendUserOperation(
+    ERC4337_BUNDLER_URL!,
     { ...unsignedUserOp, signature },
     await contracts.entryPoint.getAddress(),
   );
@@ -186,49 +143,9 @@ async function main() {
   while (!userOpReceipt) {
     // Wait 2 seconds before checking the status again
     await new Promise((resolve) => setTimeout(resolve, 2000));
-    userOpReceipt = await getUserOpReceipt(userOpHash);
+    userOpReceipt = await getUserOpReceipt(ERC4337_BUNDLER_URL!, userOpHash);
   }
   console.log("userOp Receipt", userOpReceipt);
-}
-
-/**
- * Supported Representation of UserOperation for EntryPoint v0.7
- */
-interface UserOperation {
-  sender: ethers.AddressLike;
-  nonce: string;
-  factory?: ethers.AddressLike;
-  factoryData?: ethers.BytesLike;
-  callData: string;
-  verificationGasLimit: string;
-  callGasLimit: string;
-  preVerificationGas: string;
-  maxPriorityFeePerGas: string;
-  maxFeePerGas: string;
-  signature: string;
-}
-
-export interface PaymasterResponseData {
-  paymaster?: string;
-  paymasterData?: string;
-  paymasterVerificationGasLimit?: string;
-  paymasterPostOpGasLimit?: string;
-  verificationGasLimit: string;
-  callGasLimit: string;
-  preVerificationGas: string;
-}
-
-function packPaymasterData(data: PaymasterResponseData) {
-  return data.paymaster
-    ? ethers.hexlify(
-        ethers.concat([
-          data.paymaster,
-          ethers.toBeHex(data.paymasterVerificationGasLimit || "0x", 16),
-          ethers.toBeHex(data.paymasterPostOpGasLimit || "0x", 16),
-          data.paymasterData || "0x",
-        ]),
-      )
-    : "0x";
 }
 
 main().catch((err) => {
