+++ src/near.ts
